Get rid of ValidationResponseListener
We'll have to get it out of RestController again, but the listener wasn't really helping anyhow. This should be easier after we do some work on
actions as a service...
